+++ src/main/java/tk/estecka/selfcarehive/IBeeFromNest.java
@@ -1,0 +1,6 @@
+package tk.estecka.selfcarehive;
+
+public interface IBeeFromNest {
+	void selfcarehive$markFromNest();
+	boolean selfcarehive$isFromNest();
+}